Removed ApplicationHelper patches, code moved to app/helpers/application_helper.rb and included directly from init.rb. Relates to issue #1.
